+++ Assets/Scripts/LevelGeneration/Obstacle/WallObstacle.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class WallObstacle : AbstractObstacle
 {
-  public override void Spawn(ref Platform platform, int index, int height)
+  public override void Spawn(ref Platform platform, int index, float height)
   {
     base.Spawn(ref platform, index, height);
     Transform platformT = platform.transform;
